feat: adds blocks page
